IAR-4903 Add extension function ‘ImageView.loadImage’ and replace Glide loading image codes with it
