+++ core-sample/src/main/java/com/iar/core_sample/ui/fragments/rewards/UserRewardsAdapter.kt
@@ -9,6 +9,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
 import com.iar.core_sample.R
+import com.iar.core_sample.utils.Util.loadImage
 import com.iar.iar_core.Marker
 import com.iar.iar_core.Reward
 
@@ -24,20 +25,13 @@ class UserRewardsAdapter(private val rewardsList: List<Reward>, val listener : O
     override fun onBindViewHolder(holder: RewardViewHolder, position: Int) {
         val reward = rewardsList[position]
 
-        val requestOptions = RequestOptions()
-            .override(400, 400)
         reward.image?.let{
-        Glide.with(holder.rewardImage.getContext())
-            .load(reward.image.url)
-            .placeholder(R.drawable.splash_icon)
-            .error(R.drawable.splash_icon)
-            .apply(requestOptions)
-            .into(holder.rewardImage)
+            holder.rewardImage.loadImage(it.url, holder.rewardImage.getContext())
         }
+
         holder.rewardName.text = reward.name
         holder.rewardId.text = reward.id
         holder.itemView.setOnClickListener {
-            //println("user reward clicked ${reward.id}")
             listener.onRewardItemClick(reward)
         }
     }
