+++ core-sample/src/main/java/com/iar/core_sample/utils/Util.kt
@@ -6,6 +6,9 @@ import android.util.TypedValue
 import android.view.ViewGroup
 import android.widget.EditText
 import android.widget.FrameLayout
+import android.widget.ImageView
+import com.bumptech.glide.Glide
+import com.bumptech.glide.request.RequestOptions
 import com.google.gson.Gson
 import com.iar.core_sample.R
 
@@ -14,7 +17,7 @@ object Util {
     val gson = Gson()
 
     fun setupDialogEditText(context: Context): EditText {
-        val padding =dpToPx(20, context).toInt()
+        val padding = dpToPx(20, context).toInt()
         val params = FrameLayout.LayoutParams(
             ViewGroup.LayoutParams.MATCH_PARENT,
             ViewGroup.LayoutParams.WRAP_CONTENT
@@ -34,7 +37,7 @@ object Util {
     }
 
     fun dpToPx(dp: Int, context: Context): Float {
-        val r= context.getResources()
+        val r = context.getResources()
         return TypedValue.applyDimension(
             TypedValue.COMPLEX_UNIT_DIP,
             dp.toFloat(),
@@ -42,5 +45,15 @@ object Util {
         )
     }
 
+    fun ImageView.loadImage(url: String, context: Context) {
+        val requestOptions = RequestOptions()
+            .override(400, 400)
+        Glide.with(context)
+            .load(url)
+            .placeholder(R.drawable.splash_icon)
+            .error(R.drawable.splash_icon)
+            .apply(requestOptions)
+            .into(this)
+    }
 
 }
